Fix tenant init MODPUBSUB-148
Several problems with this implementation:

1: handler.handler was called without proper response type. It should be
someting like Future.succeededFuture(PostTenantResponse.responseWith...)
Not doing this right will make RMB throw 500.

2: Liquibase and use user created always (also when module is disabled
or even purged).

3: using postTenantSync was meant to be used for UNIT TESTING and not
for production code. This means that database setup which may take
a while will be blocking. But it's not a show-stopper.

This commit just makes the post stuff to happen in loadData. This is called
on create/upgrade of a tenant. It also has a much simpler API. However,
because mod-pubsub unit tests aren't set up for async operation, the
super.postTenantSync is kept for now. It should eventually removed, so
that postTenant does not need any overriding at all.